AB#1280 Remove <script> section from index template
