--- client/platform/web-dev/settings.js
@@ -1,7 +1,0 @@
-import { loadFileAjaxSync } from "../utils"
-
-const Settings = JSON.parse(
-  loadFileAjaxSync("/api/admin/dictionary", "application/json")
-)
-
-export default Settings
