+++ client/platform/web/settings.js
@@ -1,3 +1,7 @@
-const Settings = JSON.parse(window.ANET_DATA.dictionary)
+import { loadFileAjaxSync } from "../utils"
+
+const Settings = JSON.parse(
+  loadFileAjaxSync("/api/admin/dictionary", "application/json")
+)
 
 export default Settings
